Add support for spritesheets to logo image loader
We use the JSON format standardised in the TexturePacker tool to define
frames. Currently only minimal support is implemented (no cropping or
rotating of sheet sprites.

We also use callbacks for the texture and json loading, only adding to
the images mesh when we successfully load a texture.

A new array has been added to push update functions to, to be called
during the update loop.