Animation improvements
Raise base height of logo to be above horizon. Readjust camera to center
point vertically. When splitting, zoom in more on last stage for better
corner positioning. We also raise the height when returning to main with
the controller icon bounce, for a nice effect.

Also kept some letter multiplier on first stage of lowsplit transition,
pushing to minimal during section convergence.