Refactor Pumper for ES6 async operation
The Pumper module has been rewritten with ES6 classes, reducing global
scope and moving properties into instances of the Pumper class.

Exports have been updated to define no default, but instead named
classes (just Pumper for now), and the code throughout the channels has
been updated to reflect this.

Calls to Pumper.start() in the ct-logo channel now wait for the stream
to be available before constructing the scene and configuring updates.

webrtc-adapter was added to dependencies for Pumper to use directly.